refactor(interactive.imagetool): move dialogs into new dedicated module
